The directory of the assets folder has been changed(src=>public), the files have been renamed according to the code style.
