+++ src/components/header/header.module.scss
@@ -8,11 +8,13 @@
   font-size: 16px;
   color: $color-text;
   background-color: $color-background;
+
   a {
     text-decoration: none;
     color: $color-link;
     line-height: 30px;
   }
+
   ul {
     list-style: none;
     margin: 0;
@@ -32,6 +34,7 @@
   &__info {
     display: flex;
   }
+
   &__nav {
     display: flex;
   }
@@ -39,6 +42,7 @@
 
 .notification {
   margin-right: 16px;
+
   &::before {
     width: 30px;
     content: '';
@@ -46,7 +50,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background-image: url('../../../src/assets/icons/notifications.svg');
+    background-image: url('/assets/icons/notifications.svg');
   }
 }
 
@@ -57,6 +61,7 @@
   line-height: 19px;
   color: $color-text;
   align-items: center;
+
   &::before {
     width: 30px;
     height: 30px;
@@ -65,6 +70,6 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background-image: url('../../../src/assets/icons/user_icon.svg');
+    background-image: url('/assets/icons/user_icon.svg');
   }
 }
